RD-3116 Skip audit-logs in snap-res (#3303)
* Disable audit triggers during snap-res

During snap-res, when all the data is inserted, let's disable all the
audit triggers, because the audit entries are going to be pointless
anyway and just a waste of space (they're all going to be inserted,
well, during the snap-res).

* Snapshots: omit audit-log entries

There's always going to be audit-log entries on the manager. This means
that restoring more from the snapshot, is going to always be restoring
to essentially a non-empty manager, and we're always going to run the
risk of storage id conflicts.

Unfortunate.

Well, let's not have audit-log entries in the snapshot, then.