Reorder instance relationships
Implement reordering relationships, so they can be moved around :)
If you have, in your blueprint:
    relationships:
        - target: a
          type: x
        - target: b
          type: y
and in the dep-update, you switch the order of those, then we want
to change the order in instance.relationships as well.

To do this, create a dict of reorders like {1: 0}, and prepare a new
list of relationships based on that.