test_misc: less silly assert on blueprint ids
We used to assert that the update execution's blueprint_id is the
changed blueprint, but that's so internal, man, we don't really want
to care about an execution's (!) blueprint. What does that even mean,
an execution's blueprint?

Replace those asserts with checks that we actually _do_ care about:
checking that the deployment's blueprint has changed.