RD-3271 New dep-update: fix test_modification (#3312)
* Update node relationships: correct relationship format

.source_interfaces and .target_interfaces are unnecessary and
unexpected in those entries, because they're already unfurled into
.source_operations and .target_operations, so we don't need to
submit the same data twice.

It's not like I was aware of it before, just some test asserts on the
exact shape of the relationship object and there I learned that those
need to be skipped :P

* Format relationships: correct target id

.relationship.target_id is actually the node id; just .target_id is
the instance id. It's .target_name that is the node id.
Don't ask me why, but that's how it works.

* Reorder instance relationships

Implement reordering relationships, so they can be moved around :)
If you have, in your blueprint:
    relationships:
        - target: a
          type: x
        - target: b
          type: y
and in the dep-update, you switch the order of those, then we want
to change the order in instance.relationships as well.

To do this, create a dict of reorders like {1: 0}, and prepare a new
list of relationships based on that.

* Implement reinstalling nodes.

Nodes whose properties or operations have changed, need to have their
instances reinstalled.

The reinstall function is mostly just a port of https://github.com/cloudify-cosmo/cloudify-common/blob/f93b3020fb96c66496b5635859f13a9f96075857/cloudify/plugins/workflows.py#L784-L798

* Port test_dep_up_modification for the new dep-upd workflow

Similar to all the other test files

* test_misc: less silly assert on blueprint ids

We used to assert that the update execution's blueprint_id is the
changed blueprint, but that's so internal, man, we don't really want
to care about an execution's (!) blueprint. What does that even mean,
an execution's blueprint?

Replace those asserts with checks that we actually _do_ care about:
checking that the deployment's blueprint has changed.

* Add test_dep_update_misc for the new dep-update

Not all of the test-cases just yet, but some!

* flake

* reorder: parse & validate the step separately

this is a stronger assertion on the form of the step, checking the
"nodes" and "relationships" as well

* flake, again