RD-3522 add test_deployment_update_misc for new dep-up (#3315)
* Move the pre/post-update out of the "perform" function

I want the "perform" function to only do the "install" (..uninstall,
reinstall, establish, etc) part. I'll be probably refactoring this
later again as well.

* Package all the install parameters into a class

Somewhat clarifying the interfaces around here!

So `_execute_deployment_update` is the "install" part of the workflow,
and the `InstallParams` object is the parameters to it. Everything's
been refactored to work in terms of that.

* Implement skip_install/skip_uninstall

Simple enough! Same as the old workflow, establish/unlink are also
affected by those skip_* flags

* Allow passing a custom workflow

We will run this workflow instead of the built-in install/uninstall
procedure. This is the same as using the old dep-up procedure.

To do this, also render the parameters into a dict that can be
accepted by a workflow.

* In the new workflow, skip finalize_commit

It's not needed here, and a custom workflow might attempt to run
.finalize_commit, which we need to turn into a noop, because we already
do all that stuff, in `_post_update_graph`.

* Replace this assert with a simpler one.

No need to actually look into the dep-update id. As long as it's set,
that means the custom workflow did run. It won't be set otherwise.

* test_remove_deployment: dont worry about the dep-up IDs

Just check the counts, it's enough. No need to care about the actual
IDs, anyway. If they're deleted, they're deleted.

* test_update_group: use the new workflow

- parametrize the workflow id
- skip the count check: old-style we run 2 workflows per deployment,
  new-style, we just run one. I mean, this assert isn't strictly
  necessary, we just check that the deployments were actually updated.

* test_dep_update_misc: run all tests

Run all of the tests from that file! woo

* flake

* drop object inheritance

no need in py3