restservice log_request: omit form data
We don't keep anything interesting in the form data anyway, we almost
always use json instead. (which is request.json() rather
than request.form)

However, accessing request.form (which we did for
_every single request_) breaks one important usecase: the fileserver
(and cloudify-api) auth endpoint.

For auth_request, nginx sends through all the headers, but not the
request body (because the internal auth request is a GET request).

That means the auth request gets a non-zero content-length but an
empty request body, and so accessing `request.form` waits forever
for the body. If content-type was application/json, then accessing
request.form would not wait, because in that case flask always just
returns empty for .form