cope_with_db_failover: query via session, not via the engine (#3336)
* cope_with_db_failover: query via session, not via the engine

using the engine directly leaks connections (in the tests).
Use the session so that we use the already-open connection.

Without this, to run tests (especially with -n5), I needed to start
the db with `max_connections=BIGNUMBER`. With this, the default is fine,
and indeed even small numbers are fine.

I'm not sure why did we use the engine directly. But if we do want
to use the engine directly, then we'd have to explicitly handle
closing the connection. Instead, let's just use the session, allowing
sqlalchemy to do the connection lifecycle automatically

* Also close the engine in teardown

Closes the connectionpool and all is well.