Events list: make the timestamp range end-exclusive
I'm not sure why was it working before, but now we see fails in the
test_timestamp_range test, because the test does expect exclusive
at the upper end. Were we dropping the milliseconds bit from timestamps
before?

Anyway, the range is now going to be explicitly lower-inclusive,
upper-exclusive, same as eg. python's range() function.