Events list: make the timestamp range end-exclusive (#3341)
* Events list: make the timestamp range end-exclusive

I'm not sure why was it working before, but now we see fails in the
test_timestamp_range test, because the test does expect exclusive
at the upper end. Were we dropping the milliseconds bit from timestamps
before?

Anyway, the range is now going to be explicitly lower-inclusive,
upper-exclusive, same as eg. python's range() function.

* follow-events: handle no-milliseconds timestamp format

...and in here, we don't have ms? huh. But we don't. But the :19 is
going to just cut it, and we'll handle both formats here.
Anyway, this check is only going to look for a days difference, doesn't
really care about ms.