only put the rollback in the `except` branch
oh well. All other transactions are going to have this query
pasted into them. But we can't just blanket rollback because
it breaks tests, who are doing a lot of stuff in-transaction before
actually running the test app. Bleh.

(well, not all other transactions, the first transaction in each
request)