Recursive queries for getting a tree of dependencies
IDD.get_dependencies will now run a recursive query, and return a whole
tree (ie. parents, grandparents, grand-grandparents, etc).

We use the "core" recursive query to figure out the storage-ids of
InterDepDeps that we care about, and then go from that, to selecting
either deployments, or whole InterDepDep objects.

Note: this is a generic method, not what was used in the tests; the
utility methods used there, are going to be defined in another
commit.