RD-3573 Recalculate all ancestors sub-counts&statuses
Warning: this is the big one :)

A single function that goes through all ancestors of the given
deployment(s) and recalculates their statuses.

This is using the `_get_update_tree` query, which gives back the graph
in an adjacency-list form, so we must still traverse that graph,
and sum up all the counts.

This is not used _YET_, but the point is, we'll be running this method
after any changes to the counts, instead of the current "add_to_graph",
"delete_from_graph", "recalc_after_conversion" etc etc. Just run this,
it'll fetch as much as it needs, in a single locked query, and done.

- if there's no children, just reset everything to 0 (means there once
  were children, but they were removed)
- when traversing the graph, keep counts for each parent deployment
- there's a container class for all the counts/statuses to define
  just how exactly are they to be added. It would've been easier if
  DeploymentState instances weren't just strings, but alas.
- we only fetch direct children of ancestors, so for each of those
  direct children, we just use whatever counts they declare to have
- each parent's counts are the sum of all of its children counts,
  plus 1 for every child
- this traversal needs dicts going both sides: children→parents,
  and parents→children. Unfortunate but that's how I ended up
  spelling it

Enjoy!