RD-3753 Run recalc_ancestors rather than update_exec_status
Instead of recalcing the ancestors using the old way,
in update_exec_status, run the new method.
This is just the first call-site of recalc_ancestors, but the most
important one, because this runs more often than other places

delete_deployment is going to return the affected parent ids, but
it doesn't just yet.