Rewrite test_deployment_labels_dependencies
The main idea is that creating deployments via .put_deployment
runs the whole workflow (inlined, but essentially it does), etc,
and running a test like that takes typically on the order of 4 seconds.

However those tests don't really _need_ a deployment created like that,
they can well do with just deployment objects inserted into the db,
even if they don't really have a plan or node-instances or a .yaml,
and spelled like this, those tests take on the order of .30 seconds.