Rewrite recalc_ancestors... again
This is the big one!

So the previous approach was slow as all hell. Turns out fetching
THIS MANY rows isn't very fast, and in the usual case, we'll have
a somewhat-flat hierarchy: it's common that a single environment
contains like a thousand services, and that environment is contained in
a cloud that contains a thousand environments: this gives a structure
that's only 2 deep, but two thousand wide. The previous approach
ended up selecting all two thousand rows, then (twice, actually,
because we joined on `target OR source`).

Instead, let's use the get_dependencies query we already have,
and then, update each parent separately. Yes, this is several queries.
However, it's easy to show that even if multiple queries run this at
the same time, they'll be updated in the same order: we now add an
order_by to the dependencies query!
...and experiments don't show any deadlocks either, so it seems safe.

Anyway, for each parent, we'll be now selecting the values that we
need to update it: the total counts of sub-services and sub-envs,
and their statuses.

Those NEED to be in separate queries, because parent's update, will
then of course affect the grandparent's query, because the parent
has been updated now.