fix test_validate_update_deployment_statuses_after_conversion
it's not really good, it's really in_progress

but anyway, this is just because the group-deployment is in progress.
Because it didn't have an execution _actually_ run on it.

Either way, we don't really care what's the deployment status, in
this test: we only care that the main environment's sub-status
is the same as the actual sub-deployment's status