Rewrite test_deployment_labels_dependencies (#3356)
* Rewrite test_deployment_labels_dependencies

The main idea is that creating deployments via .put_deployment
runs the whole workflow (inlined, but essentially it does), etc,
and running a test like that takes typically on the order of 4 seconds.

However those tests don't really _need_ a deployment created like that,
they can well do with just deployment objects inserted into the db,
even if they don't really have a plan or node-instances or a .yaml,
and spelled like this, those tests take on the order of .30 seconds.

* put self.user/tenant on the base class

it's now used in 2 child classes, so it goes on the parent class now.

* create_deployment: default the status to require_attn

might as well mark it require_attention before create-dep-env runs.
Null instead is silly.
The assert in `test_deployment_statuses_after_creation_without_sub_deployments`
wants this, so, sure, let's do it.

* add some tests for create-dep-env

Some basic tests for the workflow itself. This replaces a few tests
we've removed.

* when creating executions, pass the fk in

otherwise, we might try to add the execution to the session before
its workflow_id is set, which would break it.
This is due to all the `@validates` we have on the execution. A bit
unfortunate. Eventually I'd like to fix this.

* also add tenant and user to the executions

Co-authored-by: Łukasz Maksymczuk <lukasz@lukasz-maksymczuk.intra.codilime.com>