RD-3573 Rewrite label-dependency infra (#3358)
* Discard pointless dep-update tests

Those are just testing mocks. I've changed all the implementation.
Those tests are irrelevant then. I'm not adding new tests instead of
them _right now_, yet.
Anyway, this behaviour is tested by inte-tests' test_inter_deployment_dependencies.py,
test_dependencies_are_updated et al

* Make the "not found" error be a 404

* Declare is_id_unique for DeploymentDependency classes

Those models' id is a uuid, so we can declare this, to avoid having
to do the uniqueness check on a `sm.put`

* BaseDeploymentDependencies: default the id to uuid4

* get_dependencies: care about ordering

to make sure we don't get deadlocks, things need to be locked in
a stable & deterministic order;
to achieve this, we select the tree ordered by 1) depth inside the tree,
and 2) the storage_id to break ties inside levels

* Rewrite recalc_ancestors... again

This is the big one!

So the previous approach was slow as all hell. Turns out fetching
THIS MANY rows isn't very fast, and in the usual case, we'll have
a somewhat-flat hierarchy: it's common that a single environment
contains like a thousand services, and that environment is contained in
a cloud that contains a thousand environments: this gives a structure
that's only 2 deep, but two thousand wide. The previous approach
ended up selecting all two thousand rows, then (twice, actually,
because we joined on `target OR source`).

Instead, let's use the get_dependencies query we already have,
and then, update each parent separately. Yes, this is several queries.
However, it's easy to show that even if multiple queries run this at
the same time, they'll be updated in the same order: we now add an
order_by to the dependencies query!
...and experiments don't show any deadlocks either, so it seems safe.

Anyway, for each parent, we'll be now selecting the values that we
need to update it: the total counts of sub-services and sub-envs,
and their statuses.

Those NEED to be in separate queries, because parent's update, will
then of course affect the grandparent's query, because the parent
has been updated now.

* rewrite update_inter_deployment_dependencies

no need for the graph etc, we now fetch the deployments up front,
and compare with an `in`.

* drop The Big Lock

seems to work fine without it!

* recalc_ancestors calls transaction by itself now

* rewrite checking for dependency in uninstall/delete

To be honest, I'm not very sure if this is entirely correct. But
it seems to pass tests. I think this needs to be further refactored,
but for now, this will have to do.

* dep-update: set_recursive_dependencies

it's now using .get_dependents rather than the graph. Nice.

* dep_update: call .recalc_ancestors

easy enough, no need for the graph, we got the Single Method now!

* deployments.py: call recalc_ancestors

instead of all the add/remove stuff, all the methods now keep track
of what parent ids they touched, and at the end, they call
.recalc_ancestors once.

* prettify test_dep_groups & test_dld

- drop the mock tests, those dont make sense
- check for shorter error messages

* test_idd: replace the UpdateTreeTest

_get_update_tree doesn't exist anymore. The new method is WAY
simpler, it's just a query, so doesn't need that many tests.

* stop testing this mock

* update_deployment_dependencies_from_plan: use the new method

use .get_all_dependents rather than the graph

* simplify add_deployment_to_labels_graph

create the dld object, and check for cycles, don't do any graph stuff
at all.

* rename _remove_deployment_label_dependency

the other method does exactly the same thing. And this one was removed
in the previous commit. Ha.

* lowercase csys- labels by default

we do expect them to be lowercase in several places, mostly
sql queries. No reason to allow them to be uppercase.

* drop the now-unused classes

yaaaaay!!!

* flake

* fix test_validate_update_deployment_statuses_after_conversion

it's not really good, it's really in_progress

but anyway, this is just because the group-deployment is in progress.
Because it didn't have an execution _actually_ run on it.

Either way, we don't really care what's the deployment status, in
this test: we only care that the main environment's sub-status
is the same as the actual sub-deployment's status

* add a docstring

* s/does not exist/not found/

* rewrite this comment