RD-3680 Reduce memory usage when fetching labels
Fetching visibility/tenant_id like that didn't actually use the
declared relationship (the one with lazy=joined).
No, it used a separate join, and it actually emitted a cross join.
This is why the restservice starts using tens of gigabytes of ram at
some point.
To use the declared relationship, we have to use it explicitly,
like this.