RD-3663 Limit iteration in dequeueing
Rather than `while True`, make sure to limit it to a finite count.
This makes me feel safer.

Also, don't just fetch the same executions in every loop - repeating
the same and hoping for different results, is the definition of
insanity, after all.
Instead, increase offset in each iteration, to allow some crawling
through the results.

This doesn't directly fix RD-3663, but it's a stopgap for now