Query for currently-active executions
The idea is that there's going to be a small, limited amount of
currently active executions, so we can easily just query ALL of them.

Obviously: an execution that is active, is, well - currently running
in the mgmtworker, so it already uses WAY more memory than just
querying for just its id would - and there's going to be like 20, or
maybe 100 of them - not ten million.

(but there could be ten million queued executions, though)

So this method gives a report of what is currently active, split into
groups as well, for the dequeue to be able to figure out what exactly
should it run.