Exclude already-full exec-groups from dequeueing
Skip executions who are connected to an exec-group that is already
at capacity.
We figure out the group IDs python-side and just pass them in. There's
going to be a very limited amount of them only. (in degenerate cases,
up to the amount of running executions, so on the order of 100 max;
in realistic cases, there's only going to be a couple of them, like 5)