RD-3663 Dequeue more executions (#3391)
* Add a test

on master, this test fails, because we don't dequeue nearly enough
executions

* expose the exec-groups m2m table

I'm going to be using this table a lot, let's expose it
Unfortunately I have to re-declare it, because there just isn't a sane
way of pulling that declaration out of the relationship. Oh well.

* cache the get_queued_executions query (queries)

this is how you do it! Last PR that touched this, I noted that I don't
know how to properly cache a query, but now I do: db.Query(), and then
.from_statement, is how you do it.

* Query for currently-active executions

The idea is that there's going to be a small, limited amount of
currently active executions, so we can easily just query ALL of them.

Obviously: an execution that is active, is, well - currently running
in the mgmtworker, so it already uses WAY more memory than just
querying for just its id would - and there's going to be like 20, or
maybe 100 of them - not ten million.

(but there could be ten million queued executions, though)

So this method gives a report of what is currently active, split into
groups as well, for the dequeue to be able to figure out what exactly
should it run.

* Exclude already-full exec-groups from dequeueing

Skip executions who are connected to an exec-group that is already
at capacity.
We figure out the group IDs python-side and just pass them in. There's
going to be a very limited amount of them only. (in degenerate cases,
up to the amount of running executions, so on the order of 100 max;
in realistic cases, there's only going to be a couple of them, like 5)

* update the group_can_run dict

Now that we start with _some_ knowledge of execution-groups, based on
the dict returned from the _report method, we can just use that,
instead of starting from scratch with a separate dict

* Also limit the total amount!

Add a max_concurrent_workflows setting, and this is how many executions
we'll dequeue

* Use namedtuples rather than 2-item lists