RD-3861 Move dep-update tests to c_system_workflows (#3424)
* Move test_step_extraction

Move the step-extractor tests, tune them a bit to make them use
the workflow step-extractor rather than the restservice one.

The big huge json files that are used in test_all_changes_combined
should probably be removed, and the test should be made to create
the things it wants to test

RD-3860 is about rewriting that test

* Remove a debug log

Whoops. I blame the review.

* remove skipped test

* Remove other dep-update tests

Some of them should be brought back. But for right now I'm going to
just drop them.

RD-3859 is about bringing them back