dep-update workflow: take the dep-update object id as a parameter
There's 2 imaginable ways of starting a dep-update:
 1. just start the execution, and that will PUT a dep-update object
 2. create the dep-update object yourself, and that will start the
    execution

The advantage of method 1 is that it's possible to run for dep-groups,
possible to be scheduled, etc. That's why we made the separate
csys_update_deployment built-in workflow, which wraps a method-2
dep-update, in a method-1 workflow. (and we use that for groups)

The advantage of method 2 is that, well, that's what we currently have.
So, it's back-compatible. Also it's kinda a nice interface, too.

So far, for new-dep-update, we used method 1. But this is going
to be switching the approach to method 2. That is going to be
easier to make back-compatible, because of course all our old-dep-update
methods use method 2 as well.

So, the workflow must now expect that the dep-update object will
already be created. The workflow must not create it, only update.

In the future we might well switch to the other side as well, or even
allow both approaches (the workflow would then only create it
if it doesn't exist).