dep-update initiate endpoint: just start the execution
Don't use the "deployment updates manager" or nothing, just delegate
to the workflow and that's it.

The only thing we still need to do here, is parse out all the
parameters. This includes preparing the dep-update object itself.

Also I note that currently we have both /initiate, and
PUT /dep-updates/<id> that create the object; the PUT one is a bit
different, but soon it'll be good to have a single implementation.
Although the /initiate one must stay for back-compat at least.