RD-3664 update_execution_status: fail on cyclic dependencies
The post-execution IDD update might result in cyclic dependencies,
of course. In that case, let's just fail the execution. We don't
really have a way to prevent the user from creating this dependency,
we might as well tell them about it though. Before, the error was
only logged in the restservice log, and the execution seemed all A-OK.