dep-update: make sure to pass runtime_only_eval through
Make sure this attribute is set on the updated deployment.
Also now the setting is in a `if is None` basis, rather than
`new or old` basis, so it's now also possible to set the flag to False.

Even though there shouldn't really ever be a reason to do that,
but hey, it's possible.