test_idds: don't keep the might_be_deleted dependency
The test expects the "might_be_deleted" dependency to be there after
the update. But why? The property is gone from the updated blueprint.
So the idd _should_ be gone, not "might".

That it was left around in the old dep-update, was just a bug.
Now it's deleted indeed, so the assertions shouldn't expect it
to be there anymore.