augment test_modify_inputs with modifying the inputs in bp
this test will now modify the inputs, but also modifying the inputs
list in the blueprint - removes one, and adds one.

I've also made this test actually _use_ the modified blueprint at
all - you might find it funny, that the modify.yaml blueprint was
not only not used before, but it was invalid: the `test_list` input
forgot to declare the `default` key!