RD-3909 dep-up: drop removed inputs (#3440)
* augment test_modify_inputs with modifying the inputs in bp

this test will now modify the inputs, but also modifying the inputs
list in the blueprint - removes one, and adds one.

I've also made this test actually _use_ the modified blueprint at
all - you might find it funny, that the modify.yaml blueprint was
not only not used before, but it was invalid: the `test_list` input
forgot to declare the `default` key!

* RD-3909 Drop removed inputs

Inputs that were removed, need to not be passed into
prepare_deployment_plan, because passing undeclared inputs there is
just an exception.

The old inputs _will_ still be there, because even default inputs,
are copied to deployment.inputs (which is copied to dep_up.new_inputs)

* always update deployment inputs

even if the passed-in (or copied from old deployment) list is empty
(ie. old blueprint declared no inputs), update deployment inputs with
the ones from the plan - newly-declared inputs might be there.

* .get the inputs from plan

defensive!