remove some of the "unnecessary" steps
the previous commit says they're incorrect, but they're not really -
it's just, they're not necessary. They are for reasons like, the test
wants to check for something about host-agent plugins, so it adds
an operation using that plugin. The step for adding a plugin is
checked, but of course the side effect is that there's also an
"add operation" step as well, which wasn't the point, but it's there.

With the blueprints formatted like they are now, I've been able to
get rid of _some_ of the unnecessary operations, but not all