RD-3727 Dequeue executions after exiting maint-mode (#3457)
* Don't dequeue executions if we're in maintenance mode

...or activating maintenance mode.

You'd think we could allow executions to still start when activating,
or at least system workflows - at least I thought so! But alas, even
in "activating", many requests are prohibited.

* Dequeue executions when deactivating maintenance mode

Now that we don't run them when in maintenance mode, well, when are
they going to be dequeued, then? The answer is: when exiting
maintenance mode.

* add a test

in the snapshot, the execution will be queued. Is it going to be
started after a snap-res-in-maintenance? Let's find out!