RD-3748 Use tenant_name when creating IDDs (#3450)
... because of times when multiple tenants use the same deployment_id
for different deployment and we should be able to find out which one
should be used.

When creating inter-deployment dependencies, we used to look for
deployments using all_tenants=True. This way we could have used
deployments created by others (e.g. a database created by one tenant
and the application which used it by a different one). We believe however
this was not the best approach and instead suggest using globally visible
deployments it this case (so a database deployment from the example
above should be globally visible).