RD-3951 rewrite the existing_consumer_labels query
This is the important part! No longer `sm.list()`, which is subject to
pagination, and also would select labels for unrelated deployments.

That would lead it to possibly miss labels of the deployment we're
actually interested in (because they would be on the next page), which
would of course lead to adding duplicated labels, and 500s down the
road.

And if we didn't use pagination here, but just selected all of them,
well, selecting all consumer labels ever, is a bad idea, because there
might be tens of thousands of those.

Instead, only select ones based on the value as well, of which there
might still be a lot, but that's going to fit in memory for sure.

And the distinct is just an additional bonus. We're putting those in
a set anyway.