move the ifs nearer to the query
it'll allow us to exit early in some cases. Everything afterwards is
a noop anyway if the already-filtered `dependencies` is empty.

It'd be nice to be able to rewrite the `if` conditions as sql filters,
but I dont think it's possible right now, because `.external_target`
is a json. So it might well be the empty string, or the empty dict.