RD-3907 New dep-update: implement updating operation inputs
This is like #1919 (CY-1710), but for the new dep-update.
For the most part, it copies the code, just to this place.

Instead of the whole dep-update happening on the REST side, it now
happens in the workflow; but some parts must still happen in
restservice, of course.

Anyway, this is a bit heavy-handed, but I'll just move the whole update
thing to a pretty-much-RPC call in POST /operations.

This call is going to do the same thing that the old dep-up did:
examine stored operations, and update them with the new inputs.