RD-3951 update_idds: filter the labels query (#3486)
* add several tests for update_inter_deployment_dependencies

I'm going to be refactoring this a bit, so I absolutely positively
need tests.

* xfail the failing tests

those tests fail right now, because update_inter_deployment_dependencies
is buggy. I'll fix them in a later commit.

* update IDDs before filtering out components/sharedresources

make sure the IDDs are already evaluated by the time we look at
components/etc. That way, we already know the deployment id, and those
two tests now pass.

* consumer_labels_to_add: dont pass source_id separately

The labels are all pointing at the same source! This is because:
- in the shared_resources loop, we pass `deployment.id` explicitly
- in the first loop, we pass `deployment.source_deployment_id`, but
  dependencies come from a filter on _source_deployment=deployment,
  so those are ALSO always deployment.id

So no need to keep those around in the tuples, we can just pass it
through.

* RD-3951 rewrite the existing_consumer_labels query

This is the important part! No longer `sm.list()`, which is subject to
pagination, and also would select labels for unrelated deployments.

That would lead it to possibly miss labels of the deployment we're
actually interested in (because they would be on the next page), which
would of course lead to adding duplicated labels, and 500s down the
road.

And if we didn't use pagination here, but just selected all of them,
well, selecting all consumer labels ever, is a bad idea, because there
might be tens of thousands of those.

Instead, only select ones based on the value as well, of which there
might still be a lot, but that's going to fit in memory for sure.

And the distinct is just an additional bonus. We're putting those in
a set anyway.

* rewrite the list methods as sets

a bit shorter!

* move the ifs nearer to the query

it'll allow us to exit early in some cases. Everything afterwards is
a noop anyway if the already-filtered `dependencies` is empty.

It'd be nice to be able to rewrite the `if` conditions as sql filters,
but I dont think it's possible right now, because `.external_target`
is a json. So it might well be the empty string, or the empty dict.

* flake