Fixup test_custom_workflow
When using a custom workflow, the update workflow _DOES_ run now, and
calls the custom one.

This is instead of the old-dep-up idea of running the custom workflow
instead of the update workflow.

The test was passing before, because the "update" workflow wasn't
running indeed, it was named "csys-new-deployment-update". Duh.